7149 move libzpool's taskq library into libcmdutils
Reviewed by: Matt Ahrens <mahrens@delphix.com>
Reviewed by: George Wilson <george.wilson@delphix.com>

To better facilitate multi-threaded tools and libraries, it would be useful to
move the taskq implementation out of libzpool and into a more generic location
(which is currently deemed to be libcmdutils). The libzpool library can be
reworked to be a consumer of this new taskq library, and other tools/libraries
could then be made to use this library as well.

Upstream bug: DLPX-32469